requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
Levenshtein==0.27.1
multidict==6.5.0
propcache==0.3.2
RapidFuzz==3.13.0
yarl==1.20.1
